Add first controlled form with Formik
Add's a controlled form that uses the Formik library with React's
Context API to allow the user to write forms as close to HTML as
possible.

Full SSR is not yet supported.

The validation is for the full form rather than per field.

Only single page forms are supported.

Only text inputs are supported.