Forms: Clean up and prepare for forks
When we do forks we will need to gather fields from the entire graph not
simply the active path. This allows us to transform the entire graph
into a simple array of `Node`s via the new `.toArray()` method. We can
then simply gather all the fields from that.

Now that this method exists the `.deepMap()` method can piggy-back on
it in all cases so we provide an implementation for this by transforming
the `Node` interface into an abstract base class.