Add app-template package
Adds a package representing a starting point for building an application
that consumes our components using the custom Restify package.

This is very much a work in progress with lots of things that need to be
cleaned up and moved into the Restify package itself so that it does not
get in the user's way.