Site: Take components for routes and show errors
Adds support for showing any server errors that may apply to the
request.

Defaults to showing a 404 error page rather than children. (The docs
will need to be updated later.)

Routes now come in as components rather than nodes as this is neater and
more powerful. The rest of the routes should be updated to follow this
approach.