Update client renderer
Updates the client renderer to provide a HOC that will transform pages
into routes using a supplied `PageLoader` function. This cuts down on
the boiler-plate required in the applications.

The HOC is consumed by the hydrate function but also exported so that it
can be used by the server renderer. (Because of this we should rename
this package to simply `renderer` as it will be used on both the client
and the server.

Also adds the `js-enabled` class to the body tag inline with the GDS
design.