Add app-composer package
Adds an `app-composer` package for composing a full React application
from a set of components. This is designed to be shared between the
server and the client with the rendering dependant of whether the
parameters appear to have come from the server or the client. This
approach should help ensure that the both the server and client
applications are compatible with one another.