Forms: Workaround for hooks bug
For some reason the `useState` call was triggering an error relating to
hooks: https://reactjs.org/warnings/invalid-hook-call-warning.html

It seems that we don't actually need this call and so for now I have
commented it out. In the long-run we should investigate whether this
should be re-instated and the bug fixed rather than worked around.