Forms: Workaround for problem with extending Array
It seems that extending Array is not really supported. (Although this
was not a problem previously when running through CRA.) This implements
a work around that would appear to be good enough for our purposes.
(Note, however, that a downstream library wishing to extend our concept
of `Graph` might need to implement the same workaround.)

See: https://github.com/Microsoft/TypeScript/wiki/FAQ#why-doesnt-extending-built-ins-like-error-array-and-map-work