Engine: Better utilise webpack-dev-middleware
Allows the webpack-dev-middleware to write to `res.locals`. (A feature
of its `ServerSideRender` option that we need to properly link to the
CSS and JavaScript chunks in the HTML.)

Runs the middleware as a 'pre' middleware so that the new data is
available on all requests and not just requests for webpack assets.