app-template: Use node-hot-loader for local-dev
There will likely be more work required to get this working nicely but I
think the principle is sound.

We will probably need to look at how we load TypeScript in webpack and
write some `accept` functions in the libraries to support HMR. Once that
is done though we should be able to have live updates to the SSR site
without even restarting the web server.