webpack: Attempt to deal with sibling packages
The sibling packages will need to be run through webpack and so should
not be marked as external. (I've attempted to still mark their
sub-dependencies as external but am doubtful as to this works or even
could ever work.)

Also attempts to correct the config for placing the sibling packages in
their own chunk but, again, I am doubful this is having any affect.