Use pnpm's publishConfig feature
This allows us to point directly to the TypeScript code when working on
the code-base but still provide pre-transpiled code to consumers.

This should, in turn, allow us to remove our reliance on Typescript
features like 'paths' and 'project references' and allow the packages to
each have only one `tsconfig.json` file.