route-utils: Expand functionality
Provides a more comprehensive solution, including drop in replacements
for parts of react-router that provide the user with our enhanced
concept of `Location` with a pre-parsed query string.

Also provides a `queryString` helper function to build new query strings
from objects which should be useful as we will be piping a lot of state
through GET vars.

Also tightens up many of the types.