Form: Add component to docs
Unlike the others this still exists in `components` library. This is
because this component still needs the raw components in the library and
if the packages require each other we end up with infinite recursion.

Once the raw components (FormField, Button, etc...) have been migrated,
the older implementation can be removed.