create: Fix: Generate from a pre-built tarball
This allows us to more easily get special files like `.gitignore` into
place without having to worry about oddities in the `npm pack` process.
It also simplifies the generator overall.

The build process now uses `make` rather than Bash scripts which is an
extra dependency to is probably a much better fit for this sort of job.

The generator no longer runs the slow `pnpm install` process and so it
is the responsibility of the user to do so.