webpack: Fix: Avoid bundling a .node file
This seems to be attempted when `fsevents` is installed. The code
shouldn't be executed in production so I think it's okay.

On the other hand, perhaps it is wrong because it introduces the
possibility that the dependency won't be available at run-time. i.e.
Are we just denying ourselves the opportunity to fail early?