Avoid recursing into node_modules with webpack
Wildcards in the Storybook globs can cause this and our
`require.context` statements were explicitely recursive. This is a
problem as there is no way to prevent webpack from recursing into the
`node_modules/` directory which is a considerable task.

My hope is that this will speed up build times and reduce memory
consumption during the build.