Support pages loaded with a sync context
Webpack's `require.context` allows the user to set a mode. When it is
set to `sync` (the default) the module will NOT be wrapped in a promise
as it would be in `lazy` mode and some other modes.

As it is the default mode and as some users might want to set this we
should support it. To do so, we check that whether we have a promise. If
we do, we wrap the project in `lazy` as usual, otherwise there is no
need for the wrap and we can consume the component immediately.