apps: Default to sync loading of pages
This is useful as, in most cases, the extra chunks produced are so small
as to be not worth bothering with. i.e. Let's just combine them into the
main bundle.

Also, by avoid lazy loading we can avoid the need to flash up the
loading page which I am worried might be an accessibility problem.