sass-base: Have text colour cascade
This is similar to how we handle the font-family. The GDS components
set these in each components despite both the font and, in most cases,
the text colour being design dicisions that are owned elsewhere.

We get around this problem by using `inherit`. However this is a more
risky strategy with the text colour as there may be some scenarios where
`inherit` will actually bite us. As such we will have to monitor this.