route-utils: Add useIsMounted hook
Adds a hook to determing whether the component has been mounted. This is
useful in cases where we need to do something different when doing the
initial render on the server.

This does really fit very well in this package but we should probably
re-organise the packages to include a hooks or a context package.