CI: Simplify to a single workflow
This should also allow us to avoid repition of steps such as:
  - pulling dependencies
  - building applications

Deployment generators will need to be re-worked to add them as jobs to
this main workflow, rather than as workflows in their own right.

This should also allow us to run functional tests in multiple browsers
and to easily see which app has failed of passed the tests.