Split asset-proxy into its own package
BREAKING: This will require applications to update how they call the
engine.

This is helpful as it allows us to have a separate 'dev' entry for the
local-dev environment and to simplify the production build. This in turn
should allow for a simpler webpack (v5) config as well as smaller
bundles.

So far, this change seems to reduce the server 'template' bundle from
11 MiB to 5.1 MiB. (Under webpack v4.)