engine: Allow more cookies options
Allows for configuing whether cookies should be marked as 'secure'
(TLS-only). This is important as we almost always want this in
production, but not in local dev/test environments.

This also moves the `encryptionSecret` parameter into this new `cookies`
group.

Note: This would be a breaking change but we have not published the
previous interface.