Workaround to import react-helmet-async
There appears to be a bug in which named imports do not work. This bug
only appears for us when running under Jest, but not when bundling with
webpack.

This workaround allows us to import react-helmet-async in different
contexts.

If this issue is fixed upstream, then this commit can be reverted.

See: https://github.com/staylor/react-helmet-async/issues/208