BREAKING: route-utils: Abstract the router
Updates route-utils so that it abstracts the router and allows one to
work with multiple routers, as if they are react-router. Initially, this
adds support for Next.js' app router.

This is a breaking change as it removes support for features that cannot
be provided with the Next.js router, such as `useMatch`.

Removing a runtime dependence on react-router means taking some code
from react-router. It may be possible to simplify this code in the
future.