Jest: Avoid ES modules
As we are making use of require in some libraries, we now either need to
run them through a bundler like webpack, or we need to compile to
CommonJS. If we compile to ESM we get errors in Jest that require does
not exist.

Here we use a dedicated tsconfig for Jest.