template: Avoid rendering GraphQL error on server
Rendering a GraphQL error on the server currently causes problems wihen
hydrating as the client will try to load data and the 'loading' state
will not match the error state provided by the client.

By instead rendering nothing for both the error and loading states, when
the page/component is not mounted (server render and initial client
render) we make the two states identical and can side-step the issue.

It would be nice to have a better solution in the future.