app-composer: Modify interface for data routers
Modifies the interface so that we can migrate to react-router's new
'data routers'. Instead of returning an `App` component, we will return
an object containing a `RouterWrap` component and a `routes` object. It
will be the caller's (i.e. the renederer's) responisibility to set-up
the router, using both of these.

This will be a BREAKING change for anyone using just this package, but
not for anyone using the full framework, or the components.